Add preference fields to support new feature
Implementation in the tagging dialog is not yet done, but this commit
provides the two new preference (Plugin Manager) fields for
auto_select_existing_keywords, a checkbox to indicate whether we should
auto-select keywords if they exist in the keyword library, and
ignore_keyword_branches, a text area which accepts comma-separated
input for keywords which will be skipped when looking for existing
keywords, etc. This makes the plugin more performant than it might be
if it had to parse a giant list of species, etc.