Removed attempted use of debugscript
It should not have caused issues, but perhaps (possibly due to existing
Debug functionality encroaching on a namespace?) it was causing the
yield errors. Or at least commenting out and/or removing calls debug
functionality seems to resolve the issue. More tests will be run to be
sure the issue really has gone away with this change, but it seems to
be the culprit.