Almost perfect; but some issues when applying tags
Pretty close to working as it should. The dialog is built with
tooltips, etc and other improvements, but it seems that, in the end,
applying changes does not add the new keywords I’d selected. Need to
investigate, but committing this state so that (hopefully) @mcrosson
can take a look. Also note that I have commented out some calls with
table.tostring(tbl), since I was seeing a stack-overflow from it when
closing the dialog (so changes were not applied). Now, changes still
seem not to (fully) apply. Need to further investigate.